requirements.txt
einops==0.4.1
fvcore==0.1.5.post20221221
matplotlib==3.3.3
numpy==1.19.5
opencv-python==4.4.0.46
pandas==1.1.4
Pillow==8.0.1
PyYAML==5.3.1
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
tensorboardX==2.1
thop==0.1.1.post2209072238
timm==0.6.12
torch==2.5.1+cu124
torch-dct==0.1.6
torchaudio==0.8.0
torchfile==0.1.0
torchvision==0.20.1
tornado==6.1
tqdm==4.64.1
zipp==3.4.0